+++ src/renderer/app/document-context.ts
@@ -1,0 +1,4 @@
+import { rx } from "@mvui/core";
+import { WournalDocument } from "document/WournalDocument";
+
+export const DocumentCtx = new rx.Context<rx.State<WournalDocument>>();